Use typer instead of argparse
