+++ pyproject.toml
@@ -32,16 +32,13 @@ dev = [
     "pytest-cov",
     "ruff",
     "tox-direct",
+    "typer",
     "types-mock",
 ]
 
 [project.scripts]
-aa-remove-data-after = "aa_remove_data.remove_data:aa_remove_data_after"
-aa-remove-data-before = "aa_remove_data.remove_data:aa_remove_data_before"
-aa-reduce-data-by-factor = "aa_remove_data.remove_data:aa_reduce_by_factor"
-aa-reduce-data-to-period = "aa_remove_data.remove_data:aa_reduce_to_period"
-aa-print-header = "aa_remove_data.archiver_data:print_header"
-pb-2-txt = "aa_remove_data.archiver_data:pb_2_txt"
+aa-remove-data = "aa_remove_data.remove_data:app"
+pb-tools = "aa_remove_data.pb_tools:app"
 
 [project.urls]
 GitHub = "https://github.com/DiamondLightSource/aa-remove-data"
