+++ src/aa_remove_data/remove_data.py
@@ -1,310 +1,140 @@
-import argparse
 import subprocess
-from collections.abc import Iterator
 from datetime import datetime
 from pathlib import Path
-from typing import Any
 
+import typer
+
+from aa_remove_data.algorithms import (
+    apply_min_period,
+    reduce_by_factor,
+    remove_after_ts,
+    remove_before_ts,
+)
 from aa_remove_data.archiver_data import ArchiverData
 
+app = typer.Typer()
 
-def get_nano_diff(sample1: Any, sample2: Any) -> int:
-    """Get the difference in nano seconds between two samples.
+FILENAME_ARGUMENT = typer.Argument(help="path/to/file.pb of PB file being processed")
+NEW_FILENAME_OPTION = typer.Option(None, help="path/to/file.pb of new file to write to")
+BACKUP_FILENAME_OPTION = typer.Option(help="path/to/file.pb of a backup file")
+WRITE_TXT_OPTION = typer.Option(False, "--write-txt", "-t")
 
-    Args:
-        sample1 (Any): An Archiver Appliance sample.
-        sample2 (Any): Another Archiver Appliance sample.
 
-    Returns:
-        int: Difference in nanoseconds.
-    """
-    diff = (sample2.secondsintoyear - sample1.secondsintoyear) * 10**9 + (
-        sample2.nano - sample1.nano
+@app.command()
+def to_period(
+    filename: Path = FILENAME_ARGUMENT,
+    period: float = typer.Argument(help="Minimum period between each data point"),
+    new_filename: Path | None = NEW_FILENAME_OPTION,
+    backup_filename: Path | None = BACKUP_FILENAME_OPTION,
+    write_txt: bool = WRITE_TXT_OPTION,
+):
+    """Reduce the frequency of data in a PB file by setting a minimum period between
+    data points."""
+    f, new_f, backup_f = process_filenames(filename, new_filename, backup_filename)
+    if backup_f is not None:
+        subprocess.run(["cp", f, backup_f], check=True)
+
+    ad = ArchiverData(f)
+    ad.process_and_write(new_f, write_txt, apply_min_period, [period])
+
+
+@app.command()
+def by_factor(
+    filename: Path = FILENAME_ARGUMENT,
+    factor: float = typer.Argument(help="Minimum period between each data point"),
+    new_filename: Path | None = NEW_FILENAME_OPTION,
+    backup_filename: Path | None = BACKUP_FILENAME_OPTION,
+    write_txt: bool = WRITE_TXT_OPTION,
+):
+    """Reduce the number of data points in a PB file by a certain factor by removing all
+    but every nth."""
+    f, new_f, backup_f = process_filenames(filename, new_filename, backup_filename)
+    if backup_f is not None:
+        subprocess.run(["cp", f, backup_f], check=True)
+
+    ad = ArchiverData(f)
+    ad.process_and_write(
+        new_f, write_txt, reduce_by_factor, [factor], deserialize=False
     )
-    if not diff > 0:
-        raise ValueError(
-            f"diff ({diff}) is non-positive - ensure sample2 comes after sample1."
-        )
-    return diff
 
 
-def get_seconds_diff(sample1: Any, sample2: Any) -> int:
-    """Get the difference in whole seconds between two samples.
+@app.command()
+def remove_before(
+    filename: Path = FILENAME_ARGUMENT,
+    timestamp: str = typer.Argument(help="Minimum period between each data point"),
+    new_filename: Path | None = NEW_FILENAME_OPTION,
+    backup_filename: Path | None = BACKUP_FILENAME_OPTION,
+    write_txt: bool = WRITE_TXT_OPTION,
+):
+    """Remove all data points before a certain timestamp in a PB file."""
+    f, new_f, backup_f = process_filenames(filename, new_filename, backup_filename)
+    if backup_f is not None:
+        subprocess.run(["cp", f, backup_f], check=True)
 
-    Args:
-        sample1 (Any): An Archiver Appliance sample.
-        sample2 (Any): Another Archiver Appliance sample.
-
-    Returns:
-        int: Difference in seconds
-    """
-    diff = sample2.secondsintoyear - sample1.secondsintoyear
-    if not diff >= 0:
-        raise ValueError(
-            f"diff ({diff}) is negative - ensure sample2 comes after sample1."
-        )
-    return diff
-
-
-def filter_samples_to_period(samples: Iterator, period: float) -> Iterator:
-    """Reduce the frequency of a list of samples. Specify the desired minimum period.
-
-    Args:
-        samples (list): List of samples.
-        period (float): Desired minimum period between adjacent samples.
-        initial_sample (Any, optional): An initial sample to find an initial diff.
-
-    Returns:
-        list: Reduced list of samples
-    """
-    seconds_delta = period
-    nano_delta = (seconds_delta * 10**9) // 1
-    if not nano_delta >= 1:
-        raise ValueError(f"Period ({period}) must be at least 1 nanosecond.")
-
-    if seconds_delta >= 5:  # Save time for long periods by ignoring nano
-        delta = seconds_delta
-        get_diff = get_seconds_diff
-    else:
-        delta = nano_delta  # For short periods still count nano
-        get_diff = get_nano_diff
-
-    first_sample = next(samples)
-    yield first_sample
-    last_yielded_sample = first_sample
-    for sample in samples:
-        if get_diff(last_yielded_sample, sample) >= delta:
-            last_yielded_sample = sample
-            yield sample
+    ad = ArchiverData(f)
+    seconds, nano = process_timestamp(ad.header.year, timestamp)
+    ad.process_and_write(new_f, write_txt, remove_before_ts, [seconds, nano])
 
 
-def apply_min_period(samples: Iterator, period: float):
-    return (sample for sample in filter_samples_to_period(samples, period))
+@app.command()
+def remove_after(
+    filename: Path = FILENAME_ARGUMENT,
+    timestamp: str = typer.Argument(help="Minimum period between each data point"),
+    new_filename: Path | None = NEW_FILENAME_OPTION,
+    backup_filename: Path | None = BACKUP_FILENAME_OPTION,
+    write_txt: bool = WRITE_TXT_OPTION,
+):
+    """Remove all data points after a certain timestamp in a PB file."""
+    f, new_f, backup_f = process_filenames(filename, new_filename, backup_filename)
+    if backup_f is not None:
+        subprocess.run(["cp", f, backup_f], check=True)
 
-
-def is_before(sample, seconds, nano):
-    if sample.secondsintoyear < seconds or (
-        sample.secondsintoyear == seconds and sample.nano < nano
-    ):
-        return True
-    else:
-        return False
-
-
-def is_after(sample, seconds, nano):
-    if sample.secondsintoyear > seconds or (
-        sample.secondsintoyear == seconds and sample.nano > nano
-    ):
-        return True
-    else:
-        return False
-
-
-def remove_before_ts(samples: Iterator, seconds: int, nano: int = 0) -> Iterator:
-    """Remove all samples before a certain timestamp.
-
-    Args:
-        samples (list): List of samples.
-        seconds (int): Seconds portion of timestamp.
-        nano (int, optional): Nanoseconds portion of timestamp. Defaults to 0.
-
-    Returns:
-        list: Reduced list of samples.
-    """
-    if nano >= 10**9 or nano < 0:
-        seconds += nano // (10**9)
-        nano = nano % (10**9)
-    return (sample for sample in samples if not is_before(sample, seconds, nano))
-
-
-def remove_after_ts(samples: Iterator, seconds: int, nano: int = 0) -> Iterator:
-    """Remove all samples after a certain timestamp.
-
-    Args:
-        samples (list): List of samples.
-        seconds (int): Seconds portion of timestamp.
-        nano (int, optional): Nanoseconds portion of timestamp. Defaults to 0.
-
-    Returns:
-        list: Reduced list of samples.
-    """
-    if nano >= 10**9 or nano < 0:
-        seconds += nano // (10**9)
-        nano = nano % (10**9)
-    return (sample for sample in samples if not is_after(sample, seconds, nano))
-
-
-def reduce_by_factor(samples: Iterator, factor) -> Iterator:
-    """Reduce the size of a list of samples, keeping every nth sample and
-    removing the rest.
-
-    Args:
-        samples (list): List of samples
-        factor (int): Factor to reduce the data by.
-        initial (int, optional): End point of processing from a previous chunk.
-
-    Returns:
-        list: Reduced list of samples.
-    """
-    if factor < 0:
-        raise ValueError(f"Factor ({factor}) should be > 0.")
-    return (sample for i, sample in enumerate(samples) if i % factor == 0)
-
-
-def add_generic_args(parser):
-    parser.add_argument(
-        "filename", type=str, help="path/to/file.pb of PB file being processed"
-    )
-    parser.add_argument(
-        "--new-filename",
-        type=str,
-        default=None,
-        help="path/to/file.pb of new file to write to "
-        + "(default: writes over original file)",
-    )
-    parser.add_argument(
-        "--backup-filename",
-        type=str,
-        default=None,
-        help="path/to/file.pb of a backup file, "
-        + "(default: {original_filename}_backup.pb)",
-    )
-    parser.add_argument(
-        "-t",
-        "--write-txt",
-        action="store_true",
-        help="write result to a .txt file (default: False)",
-    )
-    return parser
+    ad = ArchiverData(f)
+    seconds, nano = process_timestamp(ad.header.year, timestamp)
+    ad.process_and_write(new_f, write_txt, remove_after_ts, [seconds, nano])
 
 
 def validate_pb_file(filepath, should_exist=False):
     filepath = Path(filepath)
     if filepath.suffix != ".pb":
         raise ValueError(
-            f"Invalid file extension for {filepath}: '{filepath.suffix}'. "
+            f"Invalid file extension for '{filepath}': '{filepath.suffix}'. "
             + "Expected '.pb'."
         )
-    if should_exist:
-        if not filepath.is_file():
-            raise FileNotFoundError(f"{filepath} is not a filepath.")
+    if should_exist and not filepath.is_file():
+        raise FileNotFoundError(f"No such file: '{filepath}'")
 
 
-def process_generic_args(args):
-    validate_pb_file(args.filename, should_exist=True)
+def process_filenames(f: Path, new_f: Path | None, backup_f: Path | None):
+    validate_pb_file(f, should_exist=True)
 
-    if args.backup_filename is None and (args.new_filename in (None, args.filename)):
-        args.backup_filename = args.filename.replace(".pb", "_backup.pb")
-        args.new_filename = args.filename
-    elif args.new_filename is None:
-        args.new_filename = args.filename
-    if args.backup_filename in (args.filename, args.new_filename):
+    if backup_f is None and (new_f in (None, f)):
+        backup_f = f.with_stem(f"{f.stem}_backup")
+        new_f = f
+    elif new_f is None:
+        new_f = f
+    if backup_f in (f, new_f):
         raise ValueError(
-            f"Backup filename {args.backup_filename} cannot be the same as filename or"
+            f"Backup filename {backup_f} cannot be the same as filename or"
             + " new-filename"
         )
+    validate_pb_file(new_f)
+    if backup_f is not None:
+        validate_pb_file(backup_f)
+    return f, new_f, backup_f
 
-    validate_pb_file(args.new_filename)
-    if args.backup_filename is not None:
-        validate_pb_file(args.backup_filename)
-    return args
 
-
-def process_timestamp(ad, timestamp):
-    timestamp = [int(value) for value in timestamp.split(",")]
-    if not len(timestamp) <= 6:
+def process_timestamp(year: int, timestamp: str):
+    ts = [1, 1, 0, 0, 0, 0]
+    for i, value in enumerate(timestamp.split(",")):
+        ts[i] = int(value)
+    nano = ts.pop(5)
+    if len(ts) > 5:
         raise ValueError(
             "Give timestamp in the form 'month,day,hour,minute,second,nanosecond'. "
             + "Month is required. All must be integers."
         )
-    year = ad.header.year
-    nano = timestamp.pop(5) if len(timestamp) == 6 else 0
-    diff = datetime(*([year] + timestamp)) - datetime(year, 1, 1)
+    month, day, hour, minute, second = ts
+    diff = datetime(year, month, day, hour, minute, second) - datetime(year, 1, 1)
     seconds = int(diff.total_seconds())
     return seconds, nano
-
-
-def aa_reduce_to_period():
-    """Reduce the frequency of data in a PB file by setting a minimum period between
-    data points."""
-    parser = argparse.ArgumentParser()
-    parser = add_generic_args(parser)
-    parser.add_argument(
-        "period", type=float, help="Minimum period between each data point"
-    )
-    args = parser.parse_args()
-    args = process_generic_args(args)
-
-    filename = Path(args.filename)
-    new_pb = Path(args.new_filename)
-    if args.backup_filename is not None:
-        subprocess.run(["cp", filename, Path(args.backup_filename)], check=True)
-
-    ad = ArchiverData(filename)
-    ad.process_and_write(new_pb, args.write_txt, apply_min_period, [args.period])
-
-
-def aa_reduce_by_factor():
-    """Reduce the number of data points in a PB file by a certain factor by removing all
-    but every nth."""
-    parser = argparse.ArgumentParser()
-    parser = add_generic_args(parser)
-    parser.add_argument("factor", type=int, help="factor to reduce the data by")
-    args = parser.parse_args()
-    args = process_generic_args(args)
-
-    filename = Path(args.filename)
-    new_pb = Path(args.new_filename)
-    if args.backup_filename is not None:
-        subprocess.run(["cp", filename, Path(args.backup_filename)], check=True)
-
-    ad = ArchiverData(filename)
-    ad.process_and_write(
-        new_pb, args.write_txt, reduce_by_factor, [args.factor], deserialize=False
-    )
-
-
-def aa_remove_data_before():
-    """Remove all data points before a certain timestamp in a PB file."""
-    parser = argparse.ArgumentParser()
-    parser = add_generic_args(parser)
-    parser.add_argument(
-        "timestamp",
-        type=str,
-        help="timestamp in the form 'month,day,hour,minute,second,nanosecond'"
-        + "- month is required (default: {month},1,0,0,0,0)",
-    )
-    args = parser.parse_args()
-    args = process_generic_args(args)
-
-    filename = Path(args.filename)
-    new_pb = Path(args.new_filename)
-    if args.backup_filename is not None:
-        subprocess.run(["cp", filename, Path(args.backup_filename)], check=True)
-
-    ad = ArchiverData(filename)
-    seconds, nano = process_timestamp(ad, args.timestamp)
-    ad.process_and_write(new_pb, args.write_txt, remove_before_ts, [seconds, nano])
-
-
-def aa_remove_data_after():
-    """Remove all data points after a certain timestamp in a PB file."""
-    parser = argparse.ArgumentParser()
-    parser = add_generic_args(parser)
-    parser.add_argument(
-        "timestamp",
-        type=str,
-        help="timestamp in the form 'month,day,hour,minute,second,nanosecond'"
-        + "- month is required (default: {month},1,0,0,0,0)",
-    )
-    args = parser.parse_args()
-    args = process_generic_args(args)
-
-    filename = Path(args.filename)
-    new_pb = Path(args.new_filename)
-    if args.backup_filename is not None:
-        subprocess.run(["cp", filename, Path(args.backup_filename)], check=True)
-
-    ad = ArchiverData(filename)
-    seconds, nano = process_timestamp(ad, args.timestamp)
-    ad.process_and_write(new_pb, args.write_txt, remove_after_ts, [seconds, nano])
